organize snarks into module
